Fix broken Windows build Better handling for bad string encodings Some new nullibility annotations

+++ common/main/java/com/couchbase/lite/AbstractDatabase.java
@@ -922,7 +922,7 @@ abstract class AbstractDatabase {
         synchronized (dbLock) { return getC4DatabaseLocked().nextDocExpiration(); }
     }
 
-    int purgeExpiredDocs() {
+    long purgeExpiredDocs() {
         synchronized (dbLock) { return getC4DatabaseLocked().purgeExpiredDocs(); }
     }
